requirements.txt
llama-cpp-python==0.2.56
gradio==4.19.2
